refactor : Transactional ReadOnly 설정과, RequiredArgsContstructor 어노테이션 추가
